squashed i18n=true header bug, reloacted header to its own component
