+++ webapp/src/components/forms/form-container-styles.js
@@ -2,6 +2,20 @@ import { css } from '@emotion/react'
 
 export const formContainerStyles = css`
 
+position: absolute;
+background: white;
+top: 20%;
+
+.blocker {
+  position: fixed;
+    top: 0;
+    left: 0;
+    bottom: 0;
+    right: 0;
+  background: #ffffffa3;
+}
+
+
 .transaction-form {
   display: flex;
   flex-direction: column;
